+++ packages/datasource-sql/src/types.ts
@@ -1,7 +1,18 @@
 import { Options } from 'sequelize/types';
 
 export type ConnectionOptionsObj =
-  | { uri?: string; sslMode?: SslMode } & Pick<
+  | {
+      uri?: string;
+      sslMode?: SslMode;
+      proxySocks?: {
+        userId?: string;
+        password?: string;
+        host: string;
+        port: number;
+        version?: 5;
+        command?: 'connect';
+      };
+    } & Pick<
       Options,
       | 'database'
       | 'dialect'